psp/core/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.5",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "ruff>=0.7.4",
]

[tool.ruff]
lint.select = ["I"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
